job_platform/requirements.txt
Django==5.1.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
PyYAML==6.0.1
uritemplate==4.1.1
Pillow==10.2.0
django-cors-headers==4.3.1
channels>=4.0.0
channels-redis>=4.0.0
django-filter>=24.1
django-cleanup>=8.0.0
pdfkit>=1.0.0
setuptools>=80.3.1
reportlab>=4.0.0
openai>=1.0.0